pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "clippings"
version = "0.1.0"
description = "A self-hosted, AI-powered daily news digest with configurable sources and notification channels"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "pyyaml>=6.0",
    "feedparser>=6.0.0",
    "httpx>=0.27.0",
    "litellm>=1.50.0",
    "apscheduler>=3.10.0",
    "click>=8.1.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "trafilatura>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.6.0",
]

[project.scripts]
digest = "src.main:cli"

[tool.setuptools.packages.find]
include = ["src*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
